Major update #10 #25 #26 #2
Rebuilt business internal package structure
Challenge is not an abstract class anymore
Condition is an interface
Separation between "basic" and "improve" condition sets and basic is
either an overall or a standard
Add (really) basic Pojo (variation, counter, ...) to represents each
parameter of a concept
Add TestService to check if it is possible the build every component
We can build an expression, a condition and a level entirely.
Add a postman collection file that has a set of tests. Just DL Postman
chrome extension and open test file into the app