#114 Abstract data storage
We find out that data storage and object mapping was leaking into Model
class
This commit provides a high level abstraction to store and read data
The data module is deleted and is now a domain subpackage
Data has been refactored.
Now we use MongoDBHandler. To store : pass Domain object into it, to
read passe a String and a domain object is returned
WARNING ! DATA PERSISTENCE IS CURRENTLY USED. NEW HIERARCHY HAS NOT BEEN
LINKED TO MODEL CLASS